third_party/WebKit/LayoutTests/http/tests/devtools/elements/edit/remove-node.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/elements-test.js"></script>
<script>

function removeNode(id)
{
    var child = document.getElementById(id);
    child.parentNode.removeChild(child);
}

function removeTextNode(id)
{
    document.getElementById(id).textContent = "";
}

function test() {
  var containerNode;

  TestRunner.runTestSuite([
    function testDumpInitial(next) {
      function callback(node) {
        containerNode = ElementsTestRunner.expandedNodeWithId('container');

        TestRunner.addResult('========= Original ========');
        ElementsTestRunner.dumpElementsTree(containerNode);
        next();
      }
      ElementsTestRunner.expandElementsTree(callback);
    },

    function testRemoveTextNode(next) {
      function callback() {
        TestRunner.addResult('===== Removed Text node =====');
        ElementsTestRunner.dumpElementsTree(containerNode);
        next();
      }
      TestRunner.evaluateInPage('removeTextNode(\'child1\')', callback);
    },

    function testRemoveFirst(next) {
      function callback() {
        TestRunner.addResult('===== Removed first =====');
        ElementsTestRunner.dumpElementsTree(containerNode);
        next();
      }
      TestRunner.evaluateInPage('removeNode(\'child1\')', callback);
    },

    function testRemoveMiddle(next) {
      function callback() {
        TestRunner.addResult('===== Removed middle =====');
        ElementsTestRunner.dumpElementsTree(containerNode);
        next();
      }
      TestRunner.evaluateInPage('removeNode(\'child3\')', callback);
    },

    function testRemoveLast(next) {
      function callback() {
        TestRunner.addResult('===== Removed last =====');
        ElementsTestRunner.dumpElementsTree(containerNode);
        next();
      }
      TestRunner.evaluateInPage('removeNode(\'child4\')', callback);
    },

    function testRemoveTheOnly(next) {
      function callback() {
        TestRunner.addResult('===== Removed the only =====');
        ElementsTestRunner.dumpElementsTree(containerNode);
        next();
      }
      TestRunner.evaluateInPage('removeNode(\'child2\')', callback);
    }
  ]);
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that elements panel updates dom tree structure upon node removal.
</p>

<div id="container"><div id="child1">Text</div><div id="child2"></div><div id="child3"></div><div id="child4"></div></div>

</body>
</html>
