third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/edit/remove-node-expected.txt

Tests that elements panel updates dom tree structure upon node removal.


Running: testDumpInitial
========= Original ========
- <div id="container">
      <div id="child1">Text</div>
      <div id="child2"></div>
      <div id="child3"></div>
      <div id="child4"></div>
  </div>

Running: testRemoveTextNode
===== Removed Text node =====
- <div id="container">
      <div id="child1"></div>
      <div id="child2"></div>
      <div id="child3"></div>
      <div id="child4"></div>
  </div>

Running: testRemoveFirst
===== Removed first =====
- <div id="container">
      <div id="child2"></div>
      <div id="child3"></div>
      <div id="child4"></div>
  </div>

Running: testRemoveMiddle
===== Removed middle =====
- <div id="container">
      <div id="child2"></div>
      <div id="child4"></div>
  </div>

Running: testRemoveLast
===== Removed last =====
- <div id="container">
      <div id="child2"></div>
  </div>

Running: testRemoveTheOnly
===== Removed the only =====
  <div id="container"></div>
